+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-
+## 2019-06-17
+  * changing default block id to $block['id']
+  * adding anchor support for blocks